launcher/requirements.txt
bcrypt==5.0.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
future==1.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nats-py==2.14.0
netifaces==0.11.0
protobuf==7.34.1
pyserial==3.5
PyYAML==6.0.2
toml==0.10.2
Werkzeug==3.0.4
